Separate state machine from spinner.
