+++ modules/concurrency/include/hephaestus/concurrency/spinner.h
@@ -8,30 +8,17 @@
 #include <chrono>
 #include <functional>
 #include <future>
+#include "hephaestus/concurrency/spinner_state_machine.h"
 
 namespace heph::concurrency {
 
 /// A spinner is a class that spins in a loop calling a user-defined function.
 /// If the function is blocking, the spinner will block the thread.
-/// If the input `rate_hz` is set to a non-zero value, the spinner will call the user-defined function at
-/// the given fixed rate.
+/// If the input `rate_hz` is set, the spinner will call the user-defined function at the given fixed rate.
 class Spinner {
 public:
-  using TransitionCallback = std::function<void()>;
-  using PolicyCallback = std::function<bool()>;
-
-  struct Callbacks {
-    TransitionCallback init_cb = []() {};
-    TransitionCallback spin_once_cb = []() {};
-    TransitionCallback completion_cb = []() {};
-
-    PolicyCallback shall_stop_spinning_cb = []() { return false; };  //!< Default: spin indefinitely.
-    PolicyCallback shall_restart_cb = []() { return false; };        //!< Default: do not restart.
-  };
-
-  /// @brief Create a continuous spinner which cannot be stopped by the callback.
-  /// Example: a callback which reads data from a sensor, until the spinner is stopped.
-  explicit Spinner(Callbacks&& callbacks, double rate_hz = 0);
+  /// @brief Create a continuous spinner.
+  explicit Spinner(SpinnerStateMachineCallback&& state_machine_callback, std::optional<double> rate_hz = std::nullopt);
 
   ~Spinner();
   Spinner(const Spinner&) = delete;
@@ -51,15 +38,13 @@ private:
   void spin();
 
 private:
-  enum class State : uint8_t { NOT_INITIALIZED, FAILED, READY_TO_SPIN, SPIN_SUCCESSFUL, EXIT };
-  State state_ = State::NOT_INITIALIZED;
-  SpinnerCallbacks callbacks_;
+  SpinnerStateMachineCallback state_machine_callback_;
 
   std::atomic_bool stop_requested_ = false;
   std::future<void> async_spinner_handle_;
   std::atomic_flag spinner_completed_ = ATOMIC_FLAG_INIT;
 
-  std::chrono::microseconds spin_period_;
+  std::optional<std::chrono::microseconds> spin_period_;
   std::chrono::system_clock::time_point start_timestamp_;
   std::mutex mutex_;
   std::condition_variable condition_;
